fix #19
This fixes the issue of two valid AXI requests when the address ends with "110" instruction wise. Ideally there should only be a loss of a clock or two if a cache miss occurs on an instruction ending with "110", which happens very rarely. I honestly don't see much point of `r.double_req`, and so it has just been removed entirely.